requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
ipykernel>=6.0.0
alembic==1.18.3
asttokens==3.0.1
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
greenlet==3.3.1
jedi==0.19.2
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
rich==14.3.0
scikit-learn==1.6.0
scipy==1.15.0
six==1.17.0
skforecast==0.14.0
SQLAlchemy==2.0.36
stack-data==0.6.3
statsmodels==0.14.4
threadpoolctl==3.5.0
tornado==6.5.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2025.1
wcwidth==0.2.13
platformdirs==4.3.6
pytest>=8.0.0
holidays>=0.30
pyarrow>=12.0.0
fastparquet>=2023.1.0
nbformat>=5.9.0
papermill>=2.4.0

